Remove unused searchStart page objects, and move those still in use to playground page obects
