+++ x-pack/solutions/search/test/functional_search/tests/search_onboarding_api_keys.ts
@@ -5,14 +5,12 @@
  * 2.0.
  */
 
-import expect from '@kbn/expect';
 import type { FtrProviderContext } from '../ftr_provider_context';
 
 export default function ({ getPageObjects, getService }: FtrProviderContext) {
   const pageObjects = getPageObjects([
     'searchIndexDetailsPage',
     'searchApiKeys',
-    'searchStart',
     'header',
     'common',
     'indexManagement',
@@ -50,89 +48,6 @@ export default function ({ getPageObjects, getService }: FtrProviderContext) {
       await cleanUp();
       await pageObjects.searchApiKeys.deleteAPIKeys();
       await deleteAllTestIndices();
-    });
-    describe('Elasticsearch Start [Onboarding Empty State]', () => {
-      let apiKeySession: { encoded: string; id: string } | null = null;
-      const { searchApiKeys, searchStart, searchNavigation } = pageObjects;
-      before(async () => {
-        await retry.tryWithRetries(
-          'Wait for redirect to start page',
-          async () => {
-            await pageObjects.searchStart.expectToBeOnStartPage();
-          },
-          {
-            retryCount: 2,
-            retryDelay: 1000,
-          }
-        );
-      });
-      it('should show the api key in code view', async () => {
-        await searchStart.expectToBeOnStartPage();
-        await searchStart.clickCodeViewButton();
-        // sometimes the API key exists in the cluster and its lost in sessionStorage
-        // if fails we retry to delete the API key and refresh the browser
-        await retry.try(
-          async () => {
-            await searchApiKeys.expectAPIKeyExists();
-          },
-          async () => {
-            await searchApiKeys.deleteAPIKeys();
-            await browser.refresh();
-            await searchStart.clickCodeViewButton();
-          }
-        );
-        await searchApiKeys.expectAPIKeyAvailable();
-
-        const apiKeyUI = await searchApiKeys.getAPIKeyFromUI();
-        apiKeySession = await searchApiKeys.getAPIKeyFromSessionStorage();
-
-        expect(apiKeySession).not.to.be(null);
-        expect(apiKeyUI).to.eql(apiKeySession!.encoded);
-
-        // check that when browser is refreshed, the api key is still available
-        await browser.refresh();
-        await searchStart.clickCodeViewButton();
-        await searchApiKeys.expectAPIKeyAvailable();
-        await searchStart.expectAPIKeyPreGenerated();
-        const refreshBrowserApiKeyUI = await searchApiKeys.getAPIKeyFromUI();
-        expect(refreshBrowserApiKeyUI).to.eql(apiKeyUI);
-
-        // Following tests are skipped as they are not working in the current setup. We will need to look into what is causing the issue with API key invalidation and regeneration.
-
-        // check that when api key is invalidated, a new one is generated
-        // await searchApiKeys.invalidateAPIKey(apiKeySession!.id);
-        // await browser.refresh();
-        // await searchStart.clickCodeViewButton();
-        // await searchApiKeys.expectAPIKeyAvailable();
-        // const newApiKeyUI = await searchApiKeys.getAPIKeyFromUI();
-        // expect(newApiKeyUI).to.not.eql(apiKeyUI);
-        // await searchStart.expectAPIKeyVisibleInCodeBlock(newApiKeyUI);
-      });
-
-      it('should explicitly ask to create api key when project already has an apikey', async () => {
-        await searchApiKeys.clearAPIKeySessionStorage();
-        await searchApiKeys.createAPIKey();
-        await browser.refresh();
-        await searchStart.expectToBeOnStartPage();
-        await searchStart.clickCodeViewButton();
-        await searchApiKeys.createApiKeyFromFlyout();
-        await searchApiKeys.expectAPIKeyAvailable();
-      });
-
-      it('Same API Key should be present on start page and index detail view', async () => {
-        await searchStart.clickCodeViewButton();
-        await searchApiKeys.expectAPIKeyAvailable();
-        const apiKeyUI = await searchApiKeys.getAPIKeyFromUI();
-
-        await searchStart.clickUIViewButton();
-        await searchStart.clickCreateIndexButton();
-        await searchStart.expectToBeOnIndexDetailsPage();
-
-        await searchApiKeys.expectAPIKeyAvailable();
-        const indexDetailsApiKey = await searchApiKeys.getAPIKeyFromUI();
-
-        expect(apiKeyUI).to.eql(indexDetailsApiKey);
-      });
     });
     describe('index details page', () => {
       before(async () => {
