+++ x-pack/solutions/search/test/page_objects/index.ts
@@ -10,7 +10,6 @@ import { SearchHomePageProvider } from './search_homepage';
 import { SearchIndexDetailPageProvider } from './search_index_details_page';
 import { SearchPlaygroundPageProvider } from './search_playground_page';
 import { SearchQueryRulesPageProvider } from './search_query_rules_page';
-import { SearchStartProvider } from './search_start';
 import { SearchSynonymsPageProvider } from './search_synonyms_page';
 
 export const searchSharedPageObjects = {
@@ -19,6 +18,5 @@ export const searchSharedPageObjects = {
   searchInferenceManagementPage: SearchInferenceManagementPageProvider,
   searchPlayground: SearchPlaygroundPageProvider,
   searchQueryRules: SearchQueryRulesPageProvider,
-  searchStart: SearchStartProvider,
   searchSynonyms: SearchSynonymsPageProvider,
 };
